+++ lib/omniauth/strategies/doximity_oauth2.rb
@@ -1,13 +1,18 @@
+# frozen_string_literal: true
+
 require "omniauth/strategies/oauth2"
+require "omniauth-doximity-oauth2/errors"
+require "active_support/core_ext/hash/indifferent_access"
 require "uri"
 require "rack/utils"
 require "jwt"
-require "net/http"
-require "json"
+require "faraday"
+require "multi_json"
 
 module OmniAuth
   module Strategies
-    class Doximity < OmniAuth::Strategies::OAuth2
+    # Doximity OmniAuth strategy.
+    class DoximityOauth2 < OmniAuth::Strategies::OAuth2
       DEFAULT_SCOPE = "openid profile:read:basic"
 
       option :name, "doximity"
@@ -32,9 +37,15 @@ module OmniAuth
       info do
         prune({
                 name: raw_subject_info["name"],
+                given_name: raw_subject_info["given_name"],
+                middle_name: raw_subject_info["middle_name"],
+                family_name: raw_subject_info["family_name"],
+                primary_email: raw_subject_info["primary_email"],
                 emails: raw_subject_info["emails"],
-                permissions: raw_subject_info["permissions"],
-                profile_photo_url: raw_subject_info["profile_photo_url"]
+                profile_photo_url: raw_subject_info["profile_photo_url"],
+                credentials: raw_subject_info["credentials"],
+                specialty: raw_subject_info["specialty"],
+                permissions: raw_subject_info["permissions"]
               })
       end
 
@@ -47,9 +58,9 @@ module OmniAuth
 
       credentials do
         prune({
-                access_token: raw_credential_info[:access_token],
-                refresh_token: raw_credential_info[:refresh_token],
-                expires_at: raw_credential_info[:expires_at],
+                access_token: raw_credential_info["access_token"],
+                refresh_token: raw_credential_info["refresh_token"],
+                expires_at: raw_credential_info["expires_at"],
                 scope: raw_credential_info["scope"],
                 token_type: raw_credential_info["token_type"]
               })
@@ -60,15 +71,13 @@ module OmniAuth
       end
 
       def raw_credential_info
-        @raw_credential_info ||= access_token.to_hash
+        @raw_credential_info ||= access_token.to_hash.with_indifferent_access
       end
 
       def authorize_params
         super.tap do |params|
           options[:authorize_options].each do |v|
-            if request.params[v]
-              params[v.to_sym] = request.params[v]
-            end
+            params[v.to_sym] = request.params[v] if request.params[v]
           end
 
           params[:scope] = get_scope(params)
@@ -91,8 +100,10 @@ module OmniAuth
         public_key_params = keys.find { |key| key["kid"] == header["kid"] }
         rsa_key = create_rsa_key(public_key_params["n"], public_key_params["e"])
 
-        body, _ = JWT.decode(token, rsa_key.public_key, true, { algorithm: header["alg"] })
+        body, = JWT.decode(token, rsa_key.public_key, true, { algorithm: header["alg"] })
         body
+      rescue JWT::VerificationError => e
+        raise OmniAuth::DoximityOauth2::JWTVerificationError(e, token)
       end
 
       def callback_url
@@ -108,9 +119,11 @@ module OmniAuth
 
       def request_keys
         url = options[:client_options][:site] + options[:client_options][:jwks_url]
-        uri = URI(url)
-        response = Net::HTTP.get(uri)
-        JSON.parse(response)["keys"]
+        response = Faraday.get(url)
+
+        raise OmniAuth::DoximityOauth2::JWKSRequestError(url, response) if response.status != 200
+
+        MultiJson.load(response.body)["keys"]
       end
 
       def create_rsa_key(n, e)