+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 Changelog
 =========
 
+## 1.0.0 - 05/02/2022
+  * Gem now publishes to RubyGems
+
+## 0.2.0 - 04/27/2022
+  * Change gem name
+
+## 0.1.1 - 04/18/2022
+  * Use Faraday and MultiJson
+  * Update [README.md]("./README.md")
+
+## 0.1.0 - 04/14/2022
+  * Unlock omniauth-oauth2 dependency
+
 ## 0.0.1 - 03/31/2022
   * Initialize gem